Changed all coordinates to double

+++ src/main/java/it/unibo/common/Position.java
@@ -5,6 +5,6 @@ package it.unibo.common;
  * @param x the position on the x axes.
  * @param y the position on the y axes.
  */
-public record Position(int x, int y) {
+public record Position(double x, double y) {
 
 }
